Move some things into helpers
