+++ app/controllers/upload.js
@@ -1,4 +1,5 @@
 import Ember from 'ember';
+import { showUrlDialog } from 'sharesome/helpers/show-url-dialog';
 
 export default Ember.Controller.extend({
 
@@ -21,48 +22,6 @@ export default Ember.Controller.extend({
       return "unkown";
     }
   }.property('file'),
-
-  // Has to be called from within a user action handler, e.g. 'click'
-  copyTextToClipboard(text) {
-    // Create temporary element with input value
-    // (has to be visible, so we move it out of view)
-    let tmpEl = Ember.$('<div>');
-    tmpEl.css({position: 'absolute', left: '-1000px', top: '-1000px'});
-    tmpEl.text(text);
-    Ember.$('body').append(tmpEl);
-    // Create range and selection
-    let range = document.createRange();
-    let selection = window.getSelection();
-    // Clear selection from any previous data.
-    selection.removeAllRanges();
-    // Make the range select the entire content of the contentHolder paragraph.
-    range.selectNodeContents(tmpEl.get(0));
-    // Add that range to the selection.
-    selection.addRange(range);
-    // Copy the selection to clipboard.
-    document.execCommand('copy');
-    // Clear selection if you want to.
-    selection.removeAllRanges();
-    // Remove temporary element
-    tmpEl.remove();
-  },
-
-  showUrlDialog(url) {
-    window.vex.dialog.alert(
-      `Direct URL:<p><input type="text" value="${url}"> 
-       <button class="copy-url"><img src="" alt="copy">cc</button></p>`
-    );
-    Ember.$('.vex-content input').first().select();
-
-    Ember.$('.vex-content button.copy-url').on('click', (e) => {
-      e.preventDefault();
-      this.copyTextToClipboard(Ember.$('.vex-content input').val());
-      Ember.$('.vex-content button.copy-url').html('&#10003; done');
-      setTimeout(() => {
-        window.vex.closeAll();
-      }, 1000);
-    });
-  },
 
   actions: {
 
@@ -115,7 +74,7 @@ export default Ember.Controller.extend({
           file: null,
           isUploading: false
         });
-        this.showUrlDialog(url);
+        showUrlDialog(url);
       }, error => {
         this.set('isUploading', false);
         window.vex.dialog.alert('Something bad happened during upload.<br />Please try again.');
