+++ tests/unit/helpers/show-url-dialog-test.js
@@ -1,0 +1,10 @@
+// import { showUrlDialog } from '../../../helpers/show-url-dialog';
+// import { module, test } from 'qunit';
+//
+// module('Unit | Helper | show url dialog');
+
+// TODO window.vex needs to be loaded somehow
+// test('it works', function(assert) {
+//   let result = showUrlDialog(42);
+//   assert.ok(result);
+// });
